+++ chapter-5/chapter-5.rb
@@ -10,7 +10,7 @@ puts 'Part 1:'
 puts 'hello '.+ 'world'
 puts (10.* 9).+ 9
 puts 'pig ' * 5
-# puts 5 * 'pig'
+# puts 5 * 'pig' <-- it will generate the error by ruby
 puts ''
 
 puts 'Part 2:'
@@ -68,4 +68,53 @@ puts str.ljust  lineWidth
 puts str.center lineWidth
 puts str.rjust  lineWidth
 puts str.ljust(lineWidth/2) + str.rjust(lineWidth/2)
+puts ''
+
+puts 'Part 7:'
+puts 5**2
+puts 5**0.5
+puts 7/3
+puts 7%3
+puts 365%7
+puts ''
+puts((5-2).abs)
+puts((2-5).abs)
+puts ''
+
+puts 'Part 8:'
+puts rand
+puts rand
+puts rand
+puts(rand(100))
+puts(rand(100))
+puts(rand(100))
+puts(rand(1))
+puts(rand(1))
+puts(rand(1))
+puts(rand(99999999999999999999999999999999999999999999999999999999999))
+puts('The weatherman said there is a '+rand(101).to_s+'% chance of rain,')
+puts('but you can never trust a weatherman.')
+puts ''
+srand 1776
+puts(rand(100))
+puts(rand(100))
+puts(rand(100))
+puts(rand(100))
+puts(rand(100))
+puts ''
+srand 1776
+puts(rand(100))
+puts(rand(100))
+puts(rand(100))
+puts(rand(100))
+puts(rand(100))
+puts ''
+
+puts 'Part 9:'
+puts(Math::PI)
+puts(Math::E)
+puts(Math.cos(Math::PI/3))
+puts(Math.tan(Math::PI/4))
+puts(Math.log(Math::E**2))
+puts((1 + Math.sqrt(5))/2)
 puts ''